Disable default features for arrayvec dependency
The arrayvec dependency enables its feature 'std' by default and
therefor pulls in std by default. This makes building this crate in a
no_std environment fails.

There is apparently nothing a dependent of this crate could do about it.
Features are resolved among the entire dependency tree. But we can
disable 'std' for arrayvec and leave it to others to turn it on.

See
https://doc.rust-lang.org/cargo/reference/features.html#feature-resolver-version-2
for more hints.